+++ source/B2B.Transactions.Infrastructure/Messages/TransactionIdRegistry.cs
@@ -18,25 +18,25 @@ using B2B.CimMessageAdapter.Messages;
 using B2B.Transactions.DataAccess;
 using Dapper;
 
-namespace B2B.Transactions.Infrastructure
+namespace B2B.Transactions.Infrastructure.Messages
 {
-    public class MessageIdRegistry : IMessageIds
+    public class TransactionIdRegistry : ITransactionIds
     {
         private readonly IDbConnectionFactory _connectionFactory;
 
-        public MessageIdRegistry(IDbConnectionFactory connectionFactory)
+        public TransactionIdRegistry(IDbConnectionFactory connectionFactory)
         {
             _connectionFactory = connectionFactory ?? throw new ArgumentNullException(nameof(connectionFactory));
         }
 
-        public async Task<bool> TryStoreAsync(string messageId)
+        public async Task<bool> TryStoreAsync(string transactionId)
         {
             var connection = _connectionFactory.GetOpenConnection();
 
             var result = await connection.ExecuteAsync(
-                    $"IF NOT EXISTS (SELECT * FROM dbo.MessageIds WHERE MessageId = @MessageId)" +
-                    $"INSERT INTO dbo.MessageIds(MessageId) VALUES(@MessageId)",
-                    new { MessageId = messageId })
+                    $"IF NOT EXISTS (SELECT * FROM dbo.TransactionIds WHERE TransactionId = @TransactionId)" +
+                    $"INSERT INTO dbo.TransactionIds(TransactionId) VALUES(@TransactionId)",
+                    new { TransactionId = transactionId })
                 .ConfigureAwait(false);
 
             return result == 1;
